requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.1
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appdirs==1.4.4
asyncssh==2.18.0
atpublic==5.0
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
cryptography==43.0.3
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.22.5
dvc==3.56.0
dvc-data==3.16.7
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
filelock==3.16.1
flatten-dict==0.4.2
flufl.lock==8.1.0
frozenlist==1.5.0
fsspec==2024.10.0
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
grandalf==0.8
gto==1.7.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
hydra-core==1.3.2
idna==3.10
iterative-telemetry==0.0.9
jiter==0.7.1
kombu==5.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
networkx==3.4.2
numpy==2.1.3
omegaconf==2.3.0
openai==1.54.4
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==3.11.0
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
pygit2==1.16.0
Pygments==2.18.0
pygtrie==2.5.0
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.6
scmrepo==3.3.8
semver==3.0.2
setuptools==75.4.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqltrie==0.11.1
tabulate==0.9.0
tomlkit==0.13.2
tqdm==4.67.0
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
yarl==1.17.1
zc.lockfile==3.0.post1
